requirements.txt
Deprecated==1.2.14
annotated-types==0.7.0
anyio==4.6.2post1
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
domin8==3.0.0rc1
fastapi==0.115.2
googleapis-common-protos==1.65.0
h11==0.14.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
limits==3.13.0
oauthlib==3.2.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-dbapi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-sqlite3==0.48b0
opentelemetry-instrumentation==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
packaging==24.1
protobuf==4.25.5
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.13
requests-oauthlib==2.0.0
requests==2.32.3
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.40.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wrapt==1.16.0
zipp==3.20.2
httpcore==1.0.6
httpx==0.27.2
